interfaces and constants restructured
