+++ lib/Enums.ts
@@ -1,0 +1,10 @@
+/**
+ * Metronom's log levels
+ * @enum
+ */
+export enum LogLevels {
+  Error,
+  Information,
+  All,
+  None,
+}
