+++ lib/Utilities.ts
@@ -1,6 +1,6 @@
 /* eslint-disable eqeqeq */
 import { Types } from './Constants';
-import { Schema } from './Model';
+import { Schema } from './Interfaces';
 import Logger from './Logger';
 /**
  * Utilities
@@ -101,6 +101,10 @@ export const safeWrite = async (
   return data;
 };
 
+/**
+ * Throw and log error with Metronom's internal Logger
+ * @param message {string}
+ */
 export const throwError = (message: string) => {
   Logger.error(message);
   throw new Error(message);
